+++ cookbooks/aws-parallelcluster-scheduler-plugin/recipes/init.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+#
+# Cookbook:: aws-parallelcluster-scheduler-plugin
+# Recipe:: init
+#
+# Copyright:: 2013-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with the
+# License. A copy of the License is located at
+#
+# http://aws.amazon.com/apache2.0/
+#
+# or in the "LICENSE.txt" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
+# OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
+# limitations under the License.
+
+# create system user
+include_recipe "aws-parallelcluster-scheduler-plugin::init_user"
+
+case node['cluster']['node_type']
+when 'HeadNode'
+  include_recipe 'aws-parallelcluster-scheduler-plugin::init_head_node'
+when 'ComputeFleet'
+  include_recipe 'aws-parallelcluster-scheduler-plugin::init_compute'
+else
+  raise "node_type must be HeadNode or ComputeFleet"
+end