+++ cookbooks/aws-parallelcluster-slurm/recipes/config_compute.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
 #
-# Cookbook Name:: aws-parallelcluster-slurm
+# Cookbook:: aws-parallelcluster-slurm
 # Recipe:: config_compute
 #
-# Copyright 2013-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright:: 2013-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with the
 # License. A copy of the License is located at
@@ -30,7 +30,7 @@ mount '/opt/slurm' do
   device(lazy { "#{node['cluster']['head_node_private_ip']}:/opt/slurm" })
   fstype "nfs"
   options node['cluster']['nfs']['hard_mount_options']
-  action %i[mount enable]
+  action %i(mount enable)
   retries 10
   retry_delay 6
 end