+++ cookbooks/aws-parallelcluster-test/libraries/helpers.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright:: 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License").
 # You may not use this file except in compliance with the License.