+++ cookbooks/aws-parallelcluster-test/recipes/test_imds.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
 #
-# Cookbook Name:: aws-parallelcluster
+# Cookbook:: aws-parallelcluster
 # Recipe:: test_imds
 #
-# Copyright 2013-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright:: 2013-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with the
 # License. A copy of the License is located at
@@ -42,6 +42,6 @@ allowed_users.each { |allowed_user| check_imds_access(allowed_user, true) }
 
 denied_users.each { |denied_user| check_imds_access(denied_user, false) }
 
-check_run_level_script('parallelcluster-iptables', %w[1 2 3 4 5], %w[0 6])
+check_run_level_script('parallelcluster-iptables', %w(1 2 3 4 5), %w(0 6))
 
 check_iptables_rules_file('/etc/parallelcluster/sysconfig/iptables.rules')